refactor: refatorado os inputs, e criado select field.
