+++ app/shared/components/index.ts
@@ -1,2 +1,5 @@
 export * from "./footer/index";
 export * from "./navbar/index";
+export * from "./searchHome/Search";
+export * from "./inputs/inputField";
+export * from "./inputs/maskedInput";
