+++ app/shared/components/inputs/selectField/index.tsx
@@ -1,40 +1,45 @@
 import React from "react";
 import styles from "@/app/(general)/register/register.module.scss";
 
-interface PasswordInputProps {
+interface SelectProps {
   id: string;
   name: string;
-  placeholder: string;
   required?: boolean;
   value: string;
-  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  onChange: (e: React.ChangeEvent<HTMLSelectElement>) => void;
   label: string;
+  options: { value: string; label: string }[];
 }
 
-export default function PasswordInput({
+export default function SelectField({
   id,
   name,
-  placeholder,
   required = true,
   value,
   onChange,
   label,
-}: PasswordInputProps) {
+  options,
+}: SelectProps) {
   return (
     <div className={styles.field}>
       <label htmlFor={id} className={styles.label}>
         {label}:
       </label>
-      <input
-        type="password"
+      <select
         id={id}
         name={name}
-        placeholder={placeholder}
         required={required}
         className={styles.input}
         value={value}
         onChange={onChange}
-      />
+      >
+        <option value="">Selecione</option>
+        {options.map((option) => (
+          <option key={option.value} value={option.value}>
+            {option.label}
+          </option>
+        ))}
+      </select>
     </div>
   );
 }
