+++ app/shared/components/searchHome/Search/search.module.scss
