+++ app/shared/utils/LocalStorageUtils/localStorageUtils.ts
@@ -19,14 +19,4 @@ export const localStorageUtils = {
     }
   },
 
-  getLikedPostsByUser: (userId: string): string[] => {
-    const likes = JSON.parse(localStorage.getItem("likedPosts") || "{}");
-    return likes[userId] || [];
-  },
-
-  setLikedPostsByUser: (userId: string, likedPosts: string[]): void => {
-    const likes = JSON.parse(localStorage.getItem("likedPosts") || "{}");
-    likes[userId] = likedPosts;
-    localStorage.setItem("likedPosts", JSON.stringify(likes));
-  },
 };
